Deduplicate code and add project video iFDO report generator
