+++ src/Support/Reports/Volumes/ImageAnnotations/ImageLocationReportGenerator.php
@@ -49,7 +49,9 @@ class ImageLocationReportGenerator extends AnnotationReportGenerator
             ->join('images', 'image_annotations.image_id', '=', 'images.id')
             ->join('labels', 'image_annotation_labels.label_id', '=', 'labels.id')
             ->where('images.volume_id', $this->source->id)
-            ->when($this->isRestrictedToLabels(), [$this, 'restrictToLabelsQuery'])
+            ->when($this->isRestrictedToLabels(), function ($query) {
+                return $this->restrictToLabelsQuery($query, 'image_annotation_labels');
+            })
             ->orderBy('labels.id')
             ->distinct();
