+++ tests/Support/Reports/Projects/Ifdo/ImageIfdoReportGeneratorTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Biigle\Tests\Modules\Reports\Support\Reports\Projects;
+namespace Biigle\Tests\Modules\Reports\Support\Reports\Projects\Ifdo;
 
-use Biigle\Modules\Reports\Support\Reports\Projects\ImageIfdoReportGenerator;
+use Biigle\Modules\Reports\Support\Reports\Projects\Ifdo\ImageIfdoReportGenerator;
 use Biigle\Tests\ProjectTest;
 use Biigle\Tests\VolumeTest;
 use Exception;
