+++ tests/Support/Reports/Volumes/Ifdo/VideoIfdoReportGeneratorTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Biigle\Tests\Modules\Reports\Support\Reports\Volumes;
+namespace Biigle\Tests\Modules\Reports\Support\Reports\Volumes\Ifdo;
 
 use Biigle\LabelSource;
 use Biigle\Shape;
-use Biigle\Modules\Reports\Support\Reports\Volumes\VideoIfdoReportGenerator;
+use Biigle\Modules\Reports\Support\Reports\Volumes\Ifdo\VideoIfdoReportGenerator;
 use Biigle\Modules\Reports\Volume;
 use Biigle\Tests\VideoAnnotationLabelTest;
 use Biigle\Tests\VideoAnnotationTest;
